requirements.txt
streamlit
opencv-python
imutils
Pillow
scikit-image
numpy